Refactor FindDeclarationVariableNodes and move FrameSlotAndDepth class out
