+++ src/main/java/org/truffleruby/language/locals/FrameSlotAndDepth.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2026 TruffleRuby contributors.
+ * Copyright (c) 2013-2025 Oracle and/or its affiliates.
+ * This code is released under a tri EPL/GPL/LGPL license.
+ * You can use it, redistribute it and/or modify it under the terms of the:
+ *
+ * Eclipse Public License version 2.0, or
+ * GNU General Public License version 2, or
+ * GNU Lesser General Public License version 2.1.
+ */
+package org.truffleruby.language.locals;
+
+public final class FrameSlotAndDepth {
+    public final int slot;
+    public final int depth;
+
+    public FrameSlotAndDepth(int slot, int depth) {
+        assert slot >= 0;
+        this.slot = slot;
+        this.depth = depth;
+    }
+}
